requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2
idna==3.3
oauthlib==3.2.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tweepy==4.5.0
urllib3==1.26.8
